Refactor controller into meaningful modules
