+++ controller/src/lib.rs
@@ -1,2 +1,5 @@
+mod controller;
 pub mod error;
 pub mod statemachine;
+
+pub use crate::controller::BrupopController;
